Add models of database to manager/models.py
TaskType models were created with a name field, Position with a name field, Worker which is inherited from AbstractUser with an additional Position field that links a table to another. The Task table was also created, which is connected to the TaskType table by a many-to-one relationship, with the Worker table - many-to-many.